requirements.txt
django
python-decouple
pillow
djangorestframework
markdown
django-filter
coreapi